+++ stakerservice/service.go
@@ -11,16 +11,19 @@ import (
 	"strings"
 	"sync/atomic"
 
+	"github.com/babylonlabs-io/btc-staker/metrics"
 	str "github.com/babylonlabs-io/btc-staker/staker"
 	scfg "github.com/babylonlabs-io/btc-staker/stakercfg"
 	"github.com/babylonlabs-io/btc-staker/stakerdb"
 	"github.com/btcsuite/btcd/btcec/v2"
 	"github.com/btcsuite/btcd/btcec/v2/schnorr"
 	"github.com/btcsuite/btcd/btcutil"
+	"github.com/btcsuite/btcd/chaincfg"
 	"github.com/btcsuite/btcd/chaincfg/chainhash"
 	"github.com/cometbft/cometbft/libs/log"
 	rpc "github.com/cometbft/cometbft/rpc/jsonrpc/server"
 	rpctypes "github.com/cometbft/cometbft/rpc/jsonrpc/types"
+	"go.uber.org/zap"
 
 	"github.com/lightningnetwork/lnd/kvdb"
 	"github.com/sirupsen/logrus"
@@ -43,6 +46,21 @@ type StakerService struct {
 	staker *str.App
 	logger *logrus.Logger
 	db     kvdb.Backend
+}
+
+// NewStakerServiceFromConfig creates a new staker service instance from config
+func NewStakerServiceFromConfig(
+	c *scfg.Config,
+	l *logrus.Logger,
+	z *zap.Logger,
+	db kvdb.Backend,
+	m *metrics.StakerMetrics,
+) (*StakerService, error) {
+	s, err := str.NewStakerAppFromConfig(c, l, z, db, m)
+	if err != nil {
+		return nil, fmt.Errorf("failed to create staker app: %w", err)
+	}
+	return NewStakerService(c, s, l, db), nil
 }
 
 // NewStakerService creates a new staker service instance
@@ -82,40 +100,12 @@ func (s *StakerService) stake(_ *rpctypes.Context,
 	fpBtcPks []string,
 	stakingTimeBlocks int64,
 ) (*ResultStake, error) {
-	if stakingAmount <= 0 {
-		return nil, fmt.Errorf("staking amount must be positive")
-	}
-
-	amount := btcutil.Amount(stakingAmount)
-
-	stakerAddr, err := btcutil.DecodeAddress(stakerAddress, &s.config.ActiveNetParams)
+	amount, stakerAddr, fpPubKeys, stakingTime, err := parseStkParams(stakerAddress, &s.config.ActiveNetParams, stakingAmount, fpBtcPks, stakingTimeBlocks)
 	if err != nil {
-		return nil, fmt.Errorf("error decoding staker address: %w", err)
+		return nil, err
 	}
 
-	fpPubKeys := make([]*btcec.PublicKey, 0)
-
-	for _, fpPk := range fpBtcPks {
-		fpPkBytes, err := hex.DecodeString(fpPk)
-		if err != nil {
-			return nil, fmt.Errorf("error decoding finality provider public key: %w", err)
-		}
-
-		fpSchnorrKey, err := schnorr.ParsePubKey(fpPkBytes)
-		if err != nil {
-			return nil, fmt.Errorf("error parsing finality provider public key: %w", err)
-		}
-
-		fpPubKeys = append(fpPubKeys, fpSchnorrKey)
-	}
-
-	if stakingTimeBlocks <= 0 || stakingTimeBlocks > math.MaxUint16 {
-		return nil, fmt.Errorf("staking time must be positive and lower than %d", math.MaxUint16)
-	}
-
-	stakingTimeUint16 := uint16(stakingTimeBlocks)
-
-	stakingTxHash, err := s.staker.StakeFunds(stakerAddr, amount, fpPubKeys, stakingTimeUint16)
+	stakingTxHash, err := s.staker.StakeFunds(stakerAddr, amount, fpPubKeys, stakingTime)
 	if err != nil {
 		return nil, fmt.Errorf("error staking funds: %w", err)
 	}
@@ -123,6 +113,105 @@ func (s *StakerService) stake(_ *rpctypes.Context,
 	return &ResultStake{
 		TxHash: stakingTxHash.String(),
 	}, nil
+}
+
+// stakeExpand stakes staker's requested amount of BTC
+func (s *StakerService) stakeExpand(_ *rpctypes.Context,
+	stakerAddress string,
+	stakingAmount int64,
+	fpBtcPks []string,
+	stakingTimeBlocks int64,
+	prevActiveStkTxHashHex string,
+) (*ResultStake, error) {
+	amount, stakerAddr, fpPubKeys, stakingTime, err := parseStkParams(stakerAddress, &s.config.ActiveNetParams, stakingAmount, fpBtcPks, stakingTimeBlocks)
+	if err != nil {
+		return nil, err
+	}
+
+	prevActiveStkTxHash, err := chainhash.NewHashFromStr(prevActiveStkTxHashHex)
+	if err != nil {
+		return nil, fmt.Errorf("failed to parse previous staking transaction hash hex %s: %w", prevActiveStkTxHashHex, err)
+	}
+
+	stakingTxHash, err := s.staker.StakeExpand(stakerAddr, amount, fpPubKeys, stakingTime, prevActiveStkTxHash)
+	if err != nil {
+		return nil, fmt.Errorf("error stake expand funds: %w", err)
+	}
+
+	return &ResultStake{
+		TxHash: stakingTxHash.String(),
+	}, nil
+}
+
+// consolidateUTXOs consolidates UTXOs into a single larger UTXO
+func (s *StakerService) consolidateUTXOs(_ *rpctypes.Context,
+	stakerAddress string,
+	targetAmount int64,
+) (*ResultStake, error) {
+	if targetAmount <= 0 {
+		return nil, fmt.Errorf("target amount must be positive")
+	}
+
+	stakerAddr, err := btcutil.DecodeAddress(stakerAddress, &s.config.ActiveNetParams)
+	if err != nil {
+		return nil, fmt.Errorf("error decoding staker address: %w", err)
+	}
+
+	consolidationTxHash, err := s.staker.ConsolidateUTXOs(stakerAddr, targetAmount)
+	if err != nil {
+		return nil, fmt.Errorf("error consolidating UTXOs: %w", err)
+	}
+
+	return &ResultStake{
+		TxHash: consolidationTxHash.String(),
+	}, nil
+}
+
+func parseStkParams(
+	stakerAddress string,
+	btcCfg *chaincfg.Params,
+	stakingAmount int64,
+	fpBtcPks []string,
+	stakingTimeBlocks int64,
+) (
+	amount btcutil.Amount,
+	stakerAddr btcutil.Address,
+	fpPubKeys []*btcec.PublicKey,
+	stakingTime uint16,
+	err error,
+) {
+	if stakingAmount <= 0 {
+		return amount, nil, nil, 0, fmt.Errorf("staking amount must be positive")
+	}
+
+	amount = btcutil.Amount(stakingAmount)
+
+	stakerAddr, err = btcutil.DecodeAddress(stakerAddress, btcCfg)
+	if err != nil {
+		return amount, nil, nil, 0, fmt.Errorf("error decoding staker address: %w", err)
+	}
+
+	fpPubKeys = make([]*btcec.PublicKey, 0)
+
+	for _, fpPk := range fpBtcPks {
+		fpPkBytes, err := hex.DecodeString(fpPk)
+		if err != nil {
+			return amount, nil, nil, 0, fmt.Errorf("error decoding finality provider public key: %w", err)
+		}
+
+		fpSchnorrKey, err := schnorr.ParsePubKey(fpPkBytes)
+		if err != nil {
+			return amount, nil, nil, 0, fmt.Errorf("error parsing finality provider public key: %w", err)
+		}
+
+		fpPubKeys = append(fpPubKeys, fpSchnorrKey)
+	}
+
+	if stakingTimeBlocks <= 0 || stakingTimeBlocks > math.MaxUint16 {
+		return amount, nil, nil, 0, fmt.Errorf("staking time must be positive and lower than %d", math.MaxUint16)
+	}
+
+	return amount, stakerAddr, fpPubKeys, uint16(stakingTimeBlocks), nil
 }
 
 // btcDelegationFromBtcStakingTx returns a btc delegation from a btc staking transaction
@@ -463,6 +552,8 @@ func (s *StakerService) GetRoutes() RoutesMap {
 		"health": NewRPCFunc(s.health, ""),
 		// staking API
 		"stake":                              NewRPCFunc(s.stake, "stakerAddress,stakingAmount,fpBtcPks,stakingTimeBlocks"),
+		"stake_expand":                       NewRPCFunc(s.stakeExpand, "stakerAddress,stakingAmount,fpBtcPks,stakingTimeBlocks,prevActiveStkTxHashHex"),
+		"consolidate_utxos":                  NewRPCFunc(s.consolidateUTXOs, "stakerAddress,targetAmount"),
 		"btc_delegation_from_btc_staking_tx": NewRPCFunc(s.btcDelegationFromBtcStakingTx, "stakerAddress,btcStkTxHash,covenantPksHex,covenantQuorum"),
 		"staking_details":                    NewRPCFunc(s.stakingDetails, "stakingTxHash"),
 		"spend_stake":                        NewRPCFunc(s.spendStake, "stakingTxHash"),